The "works" block class has been completed.
html - content, tags and classes were added for the "works" block. A lot of work was done to redesign the block, compared to the old version, this version is more optimized.

css - added styles and flex-element properties for block "works".

assets - added material to the assets\content-main folder to work with block "works".

@AidenOrtiz