The agency-brand project is fully completed.
Work on the "footer" block is completed.

html - the footer block has been completely redesigned to make it easier to write and improve the code in the future. Were added tags, classes and content for this block + was added personally from me block "ortiz-develop-copyright", to indicate on the site, the developer of this layout.

css - for block "footer" and its classes have been applied style properties SASS / CSS. Also flex-elements were used.

assets - ORTIZ DEVELOP team logo has been added to assets\footer folder.

js - A test development has been added, with a check of the amount of money in the user's wallet.